servers/data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Shared data layer for ANNA (MongoDB with Beanie ODM)"
authors = [
    { name = "Galih Laras Prakoso", email = "galih@braidapp.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "beanie>=1.29.0",
    "motor>=3.7.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
